+++ lib/src/main/java/com/google/cloud/datalineage/producerclient/helpers/FunctionWithException.java
@@ -1,0 +1,27 @@
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.google.cloud.datalineage.producerclient.helpers;
+
+/**
+ * Alternative for Function interface that allows to throw an exception.
+ *
+ * @param <T> function argument
+ * @param <R> returned value
+ * @param <E> exception
+ */
+@FunctionalInterface
+public interface FunctionWithException<T, R, E extends Exception> {
+  R apply(T t) throws E;
+}
